refactor(sdm): address PR feedback + tests
