+++ rust/alloy-op-evm/src/block/mod.rs
@@ -85,6 +85,10 @@ pub enum PostExecState {
         entries: Vec<SDMGasEntry>,
     },
     /// Verify canonical gas accounting using a post-exec payload embedded in the block.
+    ///
+    /// `payload` and `remaining` are not redundant: `payload` is the immutable verifier input
+    /// (kept for byte-equality comparison against the actual `0x7D` tx and for the block-number
+    /// re-check), while `remaining` is the mutable working set drained as txs are matched.
     Verifying {
         /// Decoded post-exec payload being verified.
         payload: PostExecPayload,
@@ -226,26 +230,25 @@ pub struct OpBlockExecutionCtx {
     pub post_exec_mode: PostExecMode,
 }
 
-/// Canonical gas and fee-settlement adjustment applied when a post-exec refund reduces a
-/// transaction's gas cost below the raw EVM result.
+/// Balance patch that reconciles fee distribution with the post-refund gas used.
 ///
-/// The raw EVM execution result already contains balance changes for fees charged using the raw
-/// gas used. Updating only the canonical gas used in receipts/block accounting would not rewrite
-/// those balance deltas. This carries the explicit settlement patch needed to make state match the
-/// canonical gas accounting: credit the sender back and debit the fee recipients that were paid for
-/// the refunded gas.
+/// The EVM has already paid fees out based on `evm_gas_used`, so just lowering `gas_used` in the
+/// receipt isn't enough — the sender, beneficiary, base-fee recipient, and operator-fee recipient
+/// all need their balances rolled back to match the canonical gas. This struct carries the
+/// per-recipient debits, plus the matching credit to the sender (which equals their sum).
 #[derive(Debug, Default, Clone)]
 pub struct PostExecAdjustment {
-    /// Refund amount subtracted from raw gas to produce canonical gas.
+    /// Refund amount subtracted from `evm_gas_used` to produce `canonical_gas_used`.
     pub refund: u64,
-    /// Sender balance delta to credit back.
-    pub sender_refund: U256,
-    /// Beneficiary balance delta to debit for refunded priority fees.
-    pub beneficiary_delta: U256,
-    /// Base fee recipient balance delta to debit for refunded base fees.
-    pub base_fee_delta: U256,
-    /// Operator fee recipient balance delta to debit for refunded operator fees.
-    pub operator_fee_delta: U256,
+    /// Wei to credit back to the sender (sum of the three recipient deltas below).
+    pub sender_balance_delta: U256,
+    /// Wei to debit from the block beneficiary — priority-fee share of the refund.
+    pub beneficiary_balance_delta: U256,
+    /// Wei to debit from the base-fee recipient — base-fee share of the refund.
+    pub base_fee_balance_delta: U256,
+    /// Wei to debit from the operator-fee recipient — operator-fee share of the refund
+    /// (post-Isthmus).
+    pub operator_fee_balance_delta: U256,
 }
 
 /// The result of executing an OP transaction.
@@ -259,8 +262,8 @@ pub struct OpTxResult<H, T> {
     pub is_post_exec: bool,
     /// The sender of the transaction.
     pub sender: Address,
-    /// Raw gas used returned by normal EVM execution before any canonical post-exec adjustment.
-    pub raw_gas_used: u64,
+    /// Gas used returned by normal EVM execution, before any canonical post-exec adjustment.
+    pub evm_gas_used: u64,
     /// Canonical gas used after any post-exec adjustment.
     pub canonical_gas_used: u64,
     /// Canonical post-exec adjustment, if any.
@@ -440,8 +443,11 @@ where
         tx_index: u64,
         is_deposit: bool,
         is_post_exec: bool,
-        raw_gas_used: u64,
+        evm_gas_used: u64,
     ) -> Result<u64, BlockExecutionError> {
+        // Entry-existence first: deposit and post-exec txs are called with this helper
+        // unconditionally to validate their tx index against the payload, so we can only
+        // raise the deposit/post-exec error when the payload actually targets them.
         let Some(refund) = self.post_exec.verifier_refund(tx_index) else {
             return Ok(0);
         };
@@ -458,9 +464,9 @@ where
             )));
         }
 
-        if refund > raw_gas_used {
+        if refund > evm_gas_used {
             return Err(Self::invalid_post_exec_payload(format!(
-                "payload refund {refund} exceeds raw gas used {raw_gas_used} for tx index {tx_index}"
+                "payload refund {refund} exceeds evm_gas_used {evm_gas_used} for tx index {tx_index}"
             )));
         }
 
@@ -567,31 +573,30 @@ where
 
     /// Computes the fee-settlement patch required after canonicalizing post-exec gas.
     ///
-    /// `canonical_gas_used` is enough for receipts and block gas accounting, but it is not enough
-    /// to fix state by itself: `evm.transact` has already charged the sender and paid fee
-    /// recipients according to `raw_gas_used`. The refunded gas must be translated back into the
-    /// exact balance deltas for every recipient before the state is committed.
+    /// `evm.transact` has already charged the sender and paid fee recipients according to
+    /// `evm_gas_used`. Lowering only the receipt's `gas_used` would leave those balance changes
+    /// in place. This translates the refunded gas back into the exact per-recipient deltas
+    /// `commit_transaction` then applies before state is committed.
     fn post_exec_settlement_deltas(
         &mut self,
         tx: impl RecoveredTx<R::Transaction>,
-        raw_gas_used: u64,
-        canonical_gas_used: u64,
+        evm_gas_used: u64,
+        post_exec_refund: u64,
         is_deposit: bool,
         is_post_exec: bool,
-    ) -> Result<(U256, U256, U256, U256), BlockExecutionError> {
-        if is_deposit || is_post_exec || canonical_gas_used >= raw_gas_used {
-            return Ok((U256::ZERO, U256::ZERO, U256::ZERO, U256::ZERO));
+    ) -> Result<PostExecAdjustment, BlockExecutionError> {
+        if is_deposit || is_post_exec || post_exec_refund == 0 {
+            return Ok(PostExecAdjustment::default());
         }
 
-        let gas_delta = raw_gas_used.saturating_sub(canonical_gas_used);
-        let gas_delta_u256 = U256::from(gas_delta);
+        let gas_delta_u256 = U256::from(post_exec_refund);
         let basefee = u128::from(self.evm.block().basefee());
         let spec_id = spec_by_timestamp_after_bedrock(
             &self.spec,
             self.evm.block().timestamp().saturating_to(),
         );
         let effective_gas_price = tx.tx().effective_gas_price(Some(self.evm.block().basefee()));
-        // SDM/PostExec is only enabled on forks after Isthmus, which is already post-London.
+        // SDM/PostExec is only enabled on forks after Karst, which is already post-London.
         // A saturating_sub landing at zero is intentional and consensus-valid: a legacy tx
         // with a gas price equal to the basefee pays zero priority fee, so the beneficiary
         // delta below must be zero as well — we credit back only what the beneficiary
@@ -599,45 +604,60 @@ where
         // component.
         let beneficiary_gas_price = effective_gas_price.saturating_sub(basefee);
 
-        let base_fee_delta = gas_delta_u256.saturating_mul(U256::from(basefee));
-        let beneficiary_delta = gas_delta_u256.saturating_mul(U256::from(beneficiary_gas_price));
+        let base_fee_balance_delta = gas_delta_u256.saturating_mul(U256::from(basefee));
+        let beneficiary_balance_delta =
+            gas_delta_u256.saturating_mul(U256::from(beneficiary_gas_price));
 
+        let canonical_gas_used = evm_gas_used.saturating_sub(post_exec_refund);
         let l1_block_info = self.l1_block_info(spec_id)?;
         let encoded = tx.tx().encoded_2718();
         let raw_fee =
-            l1_block_info.operator_fee_charge(encoded.as_ref(), U256::from(raw_gas_used), spec_id);
+            l1_block_info.operator_fee_charge(encoded.as_ref(), U256::from(evm_gas_used), spec_id);
         let canonical_fee = l1_block_info.operator_fee_charge(
             encoded.as_ref(),
             U256::from(canonical_gas_used),
             spec_id,
         );
-        let operator_fee_delta = raw_fee.saturating_sub(canonical_fee);
+        let operator_fee_balance_delta = raw_fee.saturating_sub(canonical_fee);
 
-        let sender_refund = gas_delta_u256
+        let sender_balance_delta = gas_delta_u256
             .saturating_mul(U256::from(effective_gas_price))
-            .saturating_add(operator_fee_delta);
+            .saturating_add(operator_fee_balance_delta);
 
-        Ok((sender_refund, beneficiary_delta, base_fee_delta, operator_fee_delta))
+        Ok(PostExecAdjustment {
+            refund: post_exec_refund,
+            sender_balance_delta,
+            beneficiary_balance_delta,
+            base_fee_balance_delta,
+            operator_fee_balance_delta,
+        })
     }
 
     fn apply_post_exec_refund_to_state(
         &mut self,
         state: &mut EvmState,
         sender: Address,
-        sender_refund: U256,
-        beneficiary_delta: U256,
-        base_fee_delta: U256,
-        operator_fee_delta: U256,
+        deltas: &PostExecAdjustment,
     ) -> Result<(), BlockExecutionError> {
         let beneficiary = self.evm.block().beneficiary();
-        Self::add_state_balance(self.evm.db_mut(), state, sender, sender_refund)?;
-        Self::sub_state_balance(self.evm.db_mut(), state, beneficiary, beneficiary_delta)?;
-        Self::sub_state_balance(self.evm.db_mut(), state, BASE_FEE_RECIPIENT, base_fee_delta)?;
+        Self::add_state_balance(self.evm.db_mut(), state, sender, deltas.sender_balance_delta)?;
+        Self::sub_state_balance(
+            self.evm.db_mut(),
+            state,
+            beneficiary,
+            deltas.beneficiary_balance_delta,
+        )?;
+        Self::sub_state_balance(
+            self.evm.db_mut(),
+            state,
+            BASE_FEE_RECIPIENT,
+            deltas.base_fee_balance_delta,
+        )?;
         Self::sub_state_balance(
             self.evm.db_mut(),
             state,
             OPERATOR_FEE_RECIPIENT,
-            operator_fee_delta,
+            deltas.operator_fee_balance_delta,
         )?;
 
         Ok(())
@@ -737,7 +757,7 @@ where
                 is_deposit: false,
                 is_post_exec: true,
                 sender: *tx.signer(),
-                raw_gas_used: 0,
+                evm_gas_used: 0,
                 canonical_gas_used: 0,
                 post_exec: None,
             });
@@ -779,40 +799,32 @@ where
             BlockExecutionError::evm(err, hash)
         })?;
 
-        let raw_gas_used = result.result.tx_gas_used();
+        let evm_gas_used = result.result.tx_gas_used();
         let post_exec_refund = if self.post_exec.is_producing() {
             let refund = self.evm.take_last_post_exec_tx_result().refund_total;
-            // The inspector's accumulated refund must never exceed the tx's raw gas. If
+            // The inspector's accumulated refund must never exceed the tx's evm_gas_used. If
             // it does, we'd emit an `SDMGasEntry` that any honest verifier would reject
-            // at pre-execution ("payload refund exceeds raw gas used"), so the sequencer
+            // at pre-execution ("payload refund exceeds evm_gas_used"), so the sequencer
             // would ship a block it can't verify itself. Fail here with a loud error
             // instead of letting `saturating_sub` mask the discrepancy.
-            if refund > raw_gas_used {
+            if refund > evm_gas_used {
                 return Err(Self::invalid_post_exec_payload(format!(
-                    "produced refund {refund} exceeds raw gas used {raw_gas_used} for tx index {tx_index}",
+                    "produced refund {refund} exceeds evm_gas_used {evm_gas_used} for tx index {tx_index}",
                 )));
             }
             refund
         } else {
-            self.verifier_post_exec_refund_for_tx(tx_index, is_deposit, false, raw_gas_used)?
+            self.verifier_post_exec_refund_for_tx(tx_index, is_deposit, false, evm_gas_used)?
         };
-        let canonical_gas_used = raw_gas_used.saturating_sub(post_exec_refund);
-        let (sender_refund, beneficiary_delta, base_fee_delta, operator_fee_delta) = self
-            .post_exec_settlement_deltas(
-                &tx,
-                raw_gas_used,
-                canonical_gas_used,
-                is_deposit,
-                false,
-            )?;
-
-        let post_exec = (post_exec_refund > 0).then_some(PostExecAdjustment {
-            refund: post_exec_refund,
-            sender_refund,
-            beneficiary_delta,
-            base_fee_delta,
-            operator_fee_delta,
-        });
+        let canonical_gas_used = evm_gas_used.saturating_sub(post_exec_refund);
+        let deltas = self.post_exec_settlement_deltas(
+            &tx,
+            evm_gas_used,
+            post_exec_refund,
+            is_deposit,
+            false,
+        )?;
+        let post_exec = (post_exec_refund > 0).then_some(deltas);
 
         Ok(OpTxResult {
             inner: EthTxResult {
@@ -823,7 +835,7 @@ where
             is_deposit,
             is_post_exec: false,
             sender: *tx.signer(),
-            raw_gas_used,
+            evm_gas_used,
             canonical_gas_used,
             post_exec,
         })
@@ -840,18 +852,13 @@ where
             is_deposit,
             is_post_exec,
             sender,
-            raw_gas_used: _,
+            evm_gas_used: _,
             canonical_gas_used,
             post_exec,
         } = output;
 
-        let PostExecAdjustment {
-            refund: post_exec_refund,
-            sender_refund,
-            beneficiary_delta,
-            base_fee_delta,
-            operator_fee_delta,
-        } = post_exec.unwrap_or_default();
+        let deltas = post_exec.unwrap_or_default();
+        let post_exec_refund = deltas.refund;
 
         if !is_deposit && !is_post_exec && post_exec_refund > 0 {
             if let Some(entries) = self.post_exec.produced_entries_mut() {
@@ -873,20 +880,13 @@ where
 
         // Canonicalizing the gas in the execution result updates receipt/block gas accounting.
         // The separate state patch below is still required because the EVM state diff already
-        // contains raw fee settlement from execution.
+        // contains the EVM-level fee settlement from execution.
         Self::canonicalize_result_gas(&mut result, post_exec_refund);
-        self.apply_post_exec_refund_to_state(
-            &mut state,
-            sender,
-            sender_refund,
-            beneficiary_delta,
-            base_fee_delta,
-            operator_fee_delta,
-        )?;
+        self.apply_post_exec_refund_to_state(&mut state, sender, &deltas)?;
 
         self.system_caller.on_state(StateChangeSource::Transaction(self.receipts.len()), &state);
 
-        // append canonical gas used
+        // add canonical gas used
         self.gas_used += canonical_gas_used;
 
         // Update DA footprint if Jovian is active
@@ -1108,579 +1108,4 @@ where
 }
 
 #[cfg(test)]
-mod tests {
-    use alloc::{string::ToString, vec};
-    use alloy_consensus::{SignableTransaction, TxLegacy, transaction::Recovered};
-    use alloy_eips::eip2718::WithEncoded;
-    use alloy_evm::{EvmEnv, ToTxEnv};
-    use alloy_hardforks::ForkCondition;
-    use alloy_op_hardforks::OpHardfork;
-    use alloy_primitives::{Address, Signature, U256, uint};
-    use op_alloy::consensus::OpTxEnvelope;
-    use op_revm::{
-        L1BlockInfo, OpBuilder, OpSpecId, OpTransaction,
-        constants::{
-            BASE_FEE_SCALAR_OFFSET, ECOTONE_L1_BLOB_BASE_FEE_SLOT, ECOTONE_L1_FEE_SCALARS_SLOT,
-            L1_BASE_FEE_SLOT, L1_BLOCK_CONTRACT, OPERATOR_FEE_SCALARS_SLOT,
-        },
-    };
-    use revm::{
-        Context, MainContext,
-        context::{BlockEnv, CfgEnv},
-        database::{CacheDB, EmptyDB, InMemoryDB, State},
-        inspector::NoOpInspector,
-        primitives::HashMap,
-        state::AccountInfo,
-    };
-
-    use crate::OpEvm;
-
-    use super::*;
-
-    #[test]
-    fn test_with_encoded() {
-        let executor_factory = OpBlockExecutorFactory::new(
-            OpAlloyReceiptBuilder::default(),
-            OpChainHardforks::op_mainnet(),
-            OpEvmFactory::<crate::OpTx>::default(),
-        );
-        let mut db = State::builder().with_database(CacheDB::<EmptyDB>::default()).build();
-        let evm = executor_factory.evm_factory.create_evm(&mut db, EvmEnv::default());
-        let mut executor = executor_factory.create_executor(evm, OpBlockExecutionCtx::default());
-        let tx = Recovered::new_unchecked(
-            OpTxEnvelope::Legacy(TxLegacy::default().into_signed(Signature::new(
-                Default::default(),
-                Default::default(),
-                Default::default(),
-            ))),
-            Address::ZERO,
-        );
-        let tx_with_encoded = WithEncoded::new(tx.encoded_2718().into(), tx.clone());
-
-        // make sure we can use both `WithEncoded` and transaction itself as inputs.
-        let _ = executor.execute_transaction(&tx);
-        let _ = executor.execute_transaction(&tx_with_encoded);
-    }
-
-    #[test]
-    fn test_settlement_state_account_preserves_original_info() {
-        type TestExecutor<'a> = OpBlockExecutor<
-            OpEvm<&'a mut State<InMemoryDB>, NoOpInspector>,
-            &'a OpAlloyReceiptBuilder,
-            &'a OpChainHardforks,
-        >;
-
-        let mut backing_db = InMemoryDB::default();
-        backing_db.insert_account_info(
-            BASE_FEE_RECIPIENT,
-            AccountInfo { balance: U256::from(10), ..Default::default() },
-        );
-        let mut db = State::builder().with_database(backing_db).with_bundle_update().build();
-        revm::Database::basic(&mut db, BASE_FEE_RECIPIENT)
-            .expect("failed to load base fee recipient into cache");
-
-        let mut credited_account =
-            Account::from(AccountInfo { balance: U256::from(15), ..Default::default() });
-        credited_account.mark_touch();
-        revm::DatabaseCommit::commit(
-            &mut db,
-            HashMap::from_iter([(BASE_FEE_RECIPIENT, credited_account)]),
-        );
-
-        let mut state = EvmState::default();
-        let mut db_ref = &mut db;
-        let account = TestExecutor::state_account_mut(&mut db_ref, &mut state, BASE_FEE_RECIPIENT)
-            .expect("failed to materialize settlement account");
-        assert_eq!(account.info.balance, U256::from(15));
-        // original_info mirrors current info here — State::commit computes the
-        // true previous value from its own cache, so the bundle stays correct.
-        assert_eq!(account.original_info.balance, U256::from(15));
-
-        account.info.balance = account.info.balance.saturating_sub(U256::from(3));
-        revm::DatabaseCommit::commit(&mut db, state);
-        db.merge_transitions(revm::database::states::bundle_state::BundleRetention::Reverts);
-
-        let bundle = db.take_bundle();
-        let bundle_account = bundle
-            .account(&BASE_FEE_RECIPIENT)
-            .expect("bundle must contain the base fee recipient");
-        assert_eq!(bundle_account.original_info.as_ref().unwrap().balance, U256::from(10));
-        assert_eq!(bundle_account.info.as_ref().unwrap().balance, U256::from(12));
-    }
-
-    fn prepare_jovian_db(da_footprint_gas_scalar: u16) -> State<InMemoryDB> {
-        const L1_BASE_FEE: U256 = uint!(1_U256);
-        const L1_BLOB_BASE_FEE: U256 = uint!(2_U256);
-        const L1_BASE_FEE_SCALAR: u64 = 3;
-        const L1_BLOB_BASE_FEE_SCALAR: u64 = 4;
-        const L1_FEE_SCALARS: U256 = U256::from_limbs([
-            0,
-            (L1_BASE_FEE_SCALAR << (64 - BASE_FEE_SCALAR_OFFSET * 2)) | L1_BLOB_BASE_FEE_SCALAR,
-            0,
-            0,
-        ]);
-        const OPERATOR_FEE_SCALAR: u8 = 5;
-        const OPERATOR_FEE_CONST: u8 = 6;
-        let da_footprint_gas_scalar_bytes = da_footprint_gas_scalar.to_be_bytes();
-        let mut operator_fee_and_da_footprint = [0u8; 32];
-        operator_fee_and_da_footprint[31] = OPERATOR_FEE_CONST;
-        operator_fee_and_da_footprint[23] = OPERATOR_FEE_SCALAR;
-        operator_fee_and_da_footprint[19] = da_footprint_gas_scalar_bytes[1];
-        operator_fee_and_da_footprint[18] = da_footprint_gas_scalar_bytes[0];
-        let operator_fee_and_da_footprint_u256 = U256::from_be_bytes(operator_fee_and_da_footprint);
-
-        let mut db = State::builder().with_database(InMemoryDB::default()).build();
-
-        db.insert_account_with_storage(
-            L1_BLOCK_CONTRACT,
-            Default::default(),
-            HashMap::from_iter([
-                (L1_BASE_FEE_SLOT, L1_BASE_FEE),
-                (ECOTONE_L1_FEE_SCALARS_SLOT, L1_FEE_SCALARS),
-                (ECOTONE_L1_BLOB_BASE_FEE_SLOT, L1_BLOB_BASE_FEE),
-                (OPERATOR_FEE_SCALARS_SLOT, operator_fee_and_da_footprint_u256),
-            ]),
-        );
-
-        db.insert_account(
-            Address::ZERO,
-            AccountInfo { balance: U256::from(400_000_000), ..Default::default() },
-        );
-
-        db
-    }
-
-    type SDMTestExecutor<'a> = OpBlockExecutor<
-        OpEvm<
-            &'a mut State<InMemoryDB>,
-            NoOpInspector,
-            op_revm::precompiles::OpPrecompiles,
-            crate::OpTx,
-        >,
-        &'a OpAlloyReceiptBuilder,
-        &'a OpChainHardforks,
-    >;
-
-    const DEFAULT_DA_FOOTPRINT_GAS_SCALAR: u16 = 7;
-    const DEFAULT_GAS_LIMIT: u64 = 100_000;
-    const JOVIAN_TIMESTAMP: u64 = 1_746_806_402;
-
-    fn build_executor<'a>(
-        db: &'a mut State<InMemoryDB>,
-        receipt_builder: &'a OpAlloyReceiptBuilder,
-        op_chain_hardforks: &'a OpChainHardforks,
-        gas_limit: u64,
-        jovian_timestamp: u64,
-    ) -> SDMTestExecutor<'a> {
-        let ctx = Context::mainnet()
-            .with_tx(crate::OpTx(OpTransaction::builder().build_fill()))
-            .with_cfg(CfgEnv::new_with_spec(OpSpecId::BEDROCK))
-            .with_chain(L1BlockInfo::default())
-            .with_db(db)
-            .with_chain(L1BlockInfo {
-                operator_fee_scalar: Some(U256::from(2)),
-                operator_fee_constant: Some(U256::from(50)),
-                ..Default::default()
-            })
-            .with_block(BlockEnv {
-                timestamp: U256::from(jovian_timestamp),
-                gas_limit,
-                ..Default::default()
-            })
-            .modify_cfg_chained(|cfg| cfg.spec = OpSpecId::JOVIAN);
-
-        let evm = OpEvm::new(ctx.build_op_with_inspector(NoOpInspector {}), true);
-
-        OpBlockExecutor::new(
-            evm,
-            OpBlockExecutionCtx::default(),
-            op_chain_hardforks,
-            receipt_builder,
-        )
-    }
-
-    struct SDMExecutorFixture {
-        db: State<InMemoryDB>,
-        receipt_builder: OpAlloyReceiptBuilder,
-        op_chain_hardforks: OpChainHardforks,
-        gas_limit: u64,
-        jovian_timestamp: u64,
-    }
-
-    impl SDMExecutorFixture {
-        fn new(da_footprint_gas_scalar: u16, gas_limit: u64, jovian_timestamp: u64) -> Self {
-            Self {
-                db: prepare_jovian_db(da_footprint_gas_scalar),
-                receipt_builder: OpAlloyReceiptBuilder::default(),
-                op_chain_hardforks: OpChainHardforks::new(
-                    OpHardfork::op_mainnet().into_iter().chain(vec![(
-                        OpHardfork::Jovian,
-                        ForkCondition::Timestamp(jovian_timestamp),
-                    )]),
-                ),
-                gas_limit,
-                jovian_timestamp,
-            }
-        }
-
-        fn default() -> Self {
-            Self::new(DEFAULT_DA_FOOTPRINT_GAS_SCALAR, DEFAULT_GAS_LIMIT, JOVIAN_TIMESTAMP)
-        }
-
-        fn executor(&mut self) -> SDMTestExecutor<'_> {
-            build_executor(
-                &mut self.db,
-                &self.receipt_builder,
-                &self.op_chain_hardforks,
-                self.gas_limit,
-                self.jovian_timestamp,
-            )
-        }
-
-        fn executor_with_post_exec_mode(
-            &mut self,
-            post_exec_mode: PostExecMode,
-        ) -> SDMTestExecutor<'_> {
-            let mut executor = self.executor();
-            executor.set_post_exec_mode(post_exec_mode);
-            executor
-        }
-    }
-
-    fn post_exec_payload(
-        block_number: u64,
-        gas_refund_entries: Vec<SDMGasEntry>,
-    ) -> PostExecPayload {
-        PostExecPayload { version: 1, block_number, gas_refund_entries }
-    }
-
-    fn legacy_tx(nonce: u64, to: Address) -> Recovered<OpTxEnvelope> {
-        let tx_inner = TxLegacy {
-            nonce,
-            gas_limit: 50_000,
-            to: alloy_primitives::TxKind::Call(to),
-            ..Default::default()
-        };
-
-        Recovered::new_unchecked(
-            OpTxEnvelope::Legacy(tx_inner.into_signed(Signature::new(
-                Default::default(),
-                Default::default(),
-                Default::default(),
-            ))),
-            Address::ZERO,
-        )
-    }
-
-    // End-to-end executor coverage for SDM: a producer emits refund entries and appends a
-    // post-exec tx, then a verifier replays the same tx stream and consumes the payload.
-    #[test]
-    fn test_post_exec_producer_verifier_roundtrip() {
-        use alloy_consensus::Sealable;
-        use op_alloy::consensus::build_post_exec_tx;
-
-        let target = Address::from([0x11; 20]);
-        let user_txs = vec![legacy_tx(0, target), legacy_tx(1, target)];
-
-        let mut producer_fixture = SDMExecutorFixture::default();
-        let mut producer = producer_fixture.executor_with_post_exec_mode(PostExecMode::Produce);
-        let first_user_gas = producer
-            .execute_transaction(&user_txs[0])
-            .expect("producer executes first user tx")
-            .tx_gas_used();
-        let second_user_gas = producer
-            .execute_transaction(&user_txs[1])
-            .expect("producer executes second user tx")
-            .tx_gas_used();
-        assert!(second_user_gas < first_user_gas, "second user tx should receive an SDM refund");
-
-        let entries = producer.take_post_exec_entries();
-        assert!(!entries.is_empty(), "producer should emit at least one SDM refund entry");
-        assert_eq!(entries[0].index, 1, "the second tx reuses block-warmed addresses");
-        assert!(entries[0].gas_refund > 0);
-
-        let post_exec_recovered = Recovered::new_unchecked(
-            OpTxEnvelope::PostExec(build_post_exec_tx(0, entries.clone()).seal_slow()),
-            Address::ZERO,
-        );
-        assert_eq!(producer.execute_transaction(&post_exec_recovered).unwrap().tx_gas_used(), 0);
-        let (_, produced) = producer.finish().expect("producer finishes block");
-
-        let mut verifier_fixture = SDMExecutorFixture::default();
-        let mut verifier = verifier_fixture
-            .executor_with_post_exec_mode(PostExecMode::Verify(post_exec_payload(0, entries)));
-        for tx in &user_txs {
-            verifier.execute_transaction(tx).expect("verifier executes user tx");
-        }
-        assert_eq!(verifier.execute_transaction(&post_exec_recovered).unwrap().tx_gas_used(), 0);
-        let (_, verified) = verifier.finish().expect("verifier consumes all entries");
-
-        assert_eq!(verified.gas_used, produced.gas_used);
-        assert_eq!(verified.blob_gas_used, produced.blob_gas_used);
-        assert_eq!(verified.receipts, produced.receipts);
-        assert_eq!(verified.receipts.len(), user_txs.len() + 1);
-    }
-
-    #[test]
-    fn test_jovian_da_footprint_estimation() {
-        let mut fixture = SDMExecutorFixture::default();
-        let mut executor = fixture.executor();
-
-        let tx_inner = TxLegacy { gas_limit: DEFAULT_GAS_LIMIT, ..Default::default() };
-
-        let tx = Recovered::new_unchecked(
-            OpTxEnvelope::Legacy(tx_inner.into_signed(Signature::new(
-                Default::default(),
-                Default::default(),
-                Default::default(),
-            ))),
-            Address::ZERO,
-        );
-        let tx_env = tx.to_tx_env();
-
-        assert!(executor.da_footprint_used == 0);
-
-        let expected_da_footprint = executor.jovian_da_footprint_estimation(&tx_env, &tx).unwrap();
-
-        // make sure we can use both `WithEncoded` and transaction itself as inputs.
-        let res = executor.execute_transaction(&tx);
-        assert!(res.is_ok());
-
-        assert!(executor.da_footprint_used == expected_da_footprint);
-    }
-
-    #[test]
-    fn test_jovian_da_footprint_estimation_out_of_gas() {
-        const GAS_LIMIT: u64 = 100;
-
-        let mut fixture =
-            SDMExecutorFixture::new(DEFAULT_DA_FOOTPRINT_GAS_SCALAR, GAS_LIMIT, JOVIAN_TIMESTAMP);
-        let mut executor = fixture.executor();
-
-        let tx_inner = TxLegacy { gas_limit: GAS_LIMIT, ..Default::default() };
-
-        let tx = Recovered::new_unchecked(
-            OpTxEnvelope::Legacy(tx_inner.into_signed(Signature::new(
-                Default::default(),
-                Default::default(),
-                Default::default(),
-            ))),
-            Address::ZERO,
-        );
-        let tx_env = tx.to_tx_env();
-
-        assert!(executor.da_footprint_used == 0);
-
-        let expected_da_footprint = executor.jovian_da_footprint_estimation(&tx_env, &tx).unwrap();
-
-        // make sure we can use both `WithEncoded` and transaction itself as inputs.
-        let res = executor.execute_transaction(&tx);
-        assert!(res.is_err());
-        let err = res.unwrap_err();
-        match err {
-            BlockExecutionError::Validation(BlockValidationError::Other(err)) => {
-                assert_eq!(
-                    err.to_string(),
-                    OpBlockExecutionError::TransactionDaFootprintAboveGasLimit {
-                        transaction_da_footprint: expected_da_footprint,
-                        available_block_da_footprint: GAS_LIMIT,
-                    }
-                    .to_string(),
-                );
-            }
-            _ => panic!("expected TransactionDaFootprintAboveGasLimit error"),
-        }
-    }
-
-    #[test]
-    fn test_jovian_da_footprint_estimation_maxed_out_da_footprint() {
-        const DA_FOOTPRINT_GAS_SCALAR: u16 = 2000;
-        const GAS_LIMIT: u64 = 200_000;
-
-        let mut fixture =
-            SDMExecutorFixture::new(DA_FOOTPRINT_GAS_SCALAR, GAS_LIMIT, JOVIAN_TIMESTAMP);
-        let mut executor = fixture.executor();
-
-        let tx_inner = TxLegacy { gas_limit: GAS_LIMIT, ..Default::default() };
-
-        let tx = Recovered::new_unchecked(
-            OpTxEnvelope::Legacy(tx_inner.into_signed(Signature::new(
-                Default::default(),
-                Default::default(),
-                Default::default(),
-            ))),
-            Address::ZERO,
-        );
-        let tx_env = tx.to_tx_env();
-
-        assert!(executor.da_footprint_used == 0);
-
-        let expected_da_footprint = executor.jovian_da_footprint_estimation(&tx_env, &tx).unwrap();
-
-        // make sure we can use both `WithEncoded` and transaction itself as inputs.
-        let gas_used_tx =
-            executor.execute_transaction(&tx).expect("failed to execute transaction").tx_gas_used();
-
-        // The gas used when executing the transaction should be the legacy value...
-        assert!(gas_used_tx < expected_da_footprint);
-
-        // The gas used when finishing the executor should be the DA footprint since this is higher
-        // than the legacy gas used and jovian is active...
-        let (_, result) = executor.finish().expect("failed to finish executor");
-        assert_eq!(result.blob_gas_used, expected_da_footprint);
-        assert_eq!(result.gas_used, gas_used_tx);
-        assert!(result.blob_gas_used > result.gas_used);
-    }
-
-    #[test]
-    fn test_mismatched_payload_block_number_fails_pre_execution() {
-        let mut fixture = SDMExecutorFixture::default();
-        // build_executor configures BlockEnv with block number 0; a payload anchored to a
-        // different block must be rejected before any tx runs.
-        let mut executor = fixture
-            .executor_with_post_exec_mode(PostExecMode::Verify(post_exec_payload(42, vec![])));
-
-        let err =
-            executor.apply_pre_execution_changes().expect_err("mismatched block number must fail");
-        assert_invalid_post_exec(err, "payload block number 42 does not match block number 0");
-    }
-
-    fn assert_invalid_post_exec(err: BlockExecutionError, expected_reason: &str) {
-        match err {
-            BlockExecutionError::Validation(BlockValidationError::Other(err)) => {
-                match err.downcast_ref::<OpBlockExecutionError>() {
-                    Some(OpBlockExecutionError::InvalidPostExecPayload(reason)) => {
-                        assert_eq!(reason, expected_reason);
-                    }
-                    other => panic!("expected invalid post-exec payload error, got: {other:?}"),
-                }
-            }
-            other => panic!("expected invalid post-exec payload error, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn test_duplicate_payload_index_fails_pre_execution() {
-        let mut fixture = SDMExecutorFixture::default();
-        // Two entries colliding on tx index 3 — the second insert must be flagged at construction
-        // and surface as a pre-execution failure.
-        let mut executor =
-            fixture.executor_with_post_exec_mode(PostExecMode::Verify(post_exec_payload(
-                0,
-                vec![
-                    SDMGasEntry { index: 3, gas_refund: 10 },
-                    SDMGasEntry { index: 3, gas_refund: 20 },
-                ],
-            )));
-
-        let err = executor
-            .apply_pre_execution_changes()
-            .expect_err("duplicate payload index must fail pre-execution");
-        assert_invalid_post_exec(err, "duplicate post-exec payload entry for tx index 3");
-    }
-
-    #[test]
-    fn test_verifier_rejects_payload_targeting_non_normal_tx() {
-        for (tx_index, is_deposit, is_post_exec, raw_gas_used, expected_reason) in [
-            (0, true, false, 21_000, "payload entry targets deposit tx index 0"),
-            (4, false, true, 0, "payload entry targets post-exec tx index 4"),
-        ] {
-            let mut fixture = SDMExecutorFixture::default();
-            let executor = fixture.executor_with_post_exec_mode(PostExecMode::Verify(
-                post_exec_payload(0, vec![SDMGasEntry { index: tx_index, gas_refund: 1 }]),
-            ));
-
-            let err = executor
-                .verifier_post_exec_refund_for_tx(tx_index, is_deposit, is_post_exec, raw_gas_used)
-                .expect_err("payload entries must not target non-normal txs");
-            assert_invalid_post_exec(err, expected_reason);
-        }
-    }
-
-    #[test]
-    fn test_verifier_rejects_refund_exceeding_raw_gas() {
-        let mut fixture = SDMExecutorFixture::default();
-        let executor = fixture.executor_with_post_exec_mode(PostExecMode::Verify(
-            post_exec_payload(0, vec![SDMGasEntry { index: 2, gas_refund: 50_000 }]),
-        ));
-
-        // raw_gas_used < payload refund — a refund that exceeds the tx's raw cost is
-        // impossible under SDM semantics and must be rejected, otherwise canonical gas
-        // would underflow to a bogus value via saturating_sub.
-        let err = executor
-            .verifier_post_exec_refund_for_tx(2, false, false, 40_000)
-            .expect_err("refund greater than raw gas must be rejected");
-        assert_invalid_post_exec(
-            err,
-            "payload refund 50000 exceeds raw gas used 40000 for tx index 2",
-        );
-
-        // Boundary: refund == raw_gas_used is permitted (canonical gas ends up at zero).
-        let ok = executor
-            .verifier_post_exec_refund_for_tx(2, false, false, 50_000)
-            .expect("refund equal to raw gas is permitted");
-        assert_eq!(ok, 50_000);
-    }
-
-    #[test]
-    fn test_verifier_returns_zero_when_no_entry_for_tx() {
-        let mut fixture = SDMExecutorFixture::default();
-        let executor = fixture.executor_with_post_exec_mode(PostExecMode::Verify(
-            post_exec_payload(0, vec![SDMGasEntry { index: 7, gas_refund: 42 }]),
-        ));
-
-        // Normal tx that has no entry in the payload — the deposit/post-exec guards must NOT
-        // fire, the helper must return 0 so execution proceeds with raw gas unchanged.
-        let refund = executor
-            .verifier_post_exec_refund_for_tx(3, false, false, 21_000)
-            .expect("no entry for this tx index means no refund");
-        assert_eq!(refund, 0);
-    }
-
-    #[test]
-    fn test_finish_reports_all_unconsumed_post_exec_entries() {
-        let mut fixture = SDMExecutorFixture::default();
-        let executor =
-            fixture.executor_with_post_exec_mode(PostExecMode::Verify(post_exec_payload(
-                0,
-                vec![
-                    SDMGasEntry { index: 2, gas_refund: 7 },
-                    SDMGasEntry { index: 5, gas_refund: 11 },
-                ],
-            )));
-
-        let Err(err) = executor.finish() else {
-            panic!("unconsumed verifier entries must fail");
-        };
-        assert_invalid_post_exec(
-            err,
-            "2 unconsumed post-exec payload entries for tx indexes [2, 5]",
-        );
-    }
-
-    /// Followers running with SDM disabled must reject any block that carries a post-exec
-    /// 0x7D tx. Silently short-circuiting the tx (which is what the pre-guard code did) would
-    /// let a producer ship a payload with arbitrary refund entries that no follower validates,
-    /// and the two nodes' states would diverge without anyone noticing.
-    #[test]
-    fn test_disabled_mode_rejects_post_exec_tx() {
-        use alloy_consensus::Sealable;
-        use op_alloy::consensus::build_post_exec_tx;
-
-        let mut fixture = SDMExecutorFixture::default();
-        // build_executor leaves post_exec_mode at the default (Disabled).
-        let mut executor = fixture.executor();
-        assert!(matches!(executor.post_exec, PostExecState::Disabled));
-
-        let post_exec_tx = build_post_exec_tx(0, vec![]);
-        let envelope = OpTxEnvelope::PostExec(post_exec_tx.seal_slow());
-        let tx = Recovered::new_unchecked(envelope, Address::ZERO);
-
-        let err =
-            executor.execute_transaction(&tx).expect_err("0x7D tx in Disabled mode must fail");
-        assert_invalid_post_exec(
-            err,
-            "unexpected post-exec tx at index 0: SDM not active for this block",
-        );
-    }
-}
+mod tests;
